sam3-labeling-tool/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.26.0
opencv-python>=4.8.0
pydantic>=2.5.0
aiofiles>=23.2.0
